+++ fbpic/lpa_utils/laser/direct_envelope_injection.py
@@ -34,6 +34,11 @@ def add_laser_direct_envelope( sim, laser_profile, boost ):
     if abs(laser_wavelength - sim.fld.lambda_envelope) > rtol*laser_wavelength:
         raise ValueError("Wavelengths passed to the simulation and the "
                     "laser profile are different")
+    # Check that the propagation direction is correct
+    if not laser_profile.propag_direction == 1.:
+        raise ValueError(
+            "The envelope model only works for forward-propagating lasers.\n"
+            "Set the `propagation_direction` argument to 1.")
 
     # Get the local azimuthally-decomposed laser fields a and a_old on each proc
     laser_a, laser_a_old = get_laser_a( sim, laser_profile, boost )